# api/openapi.yaml
openapi: 3.0.3
info:
  title: SurveyFlow API
  description: SurveyFlow backend API for creating, distributing, and analyzing surveys.
  version: 1.0.0
  contact:
    name: SurveyFlow
    email: api@surveyflow.io
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/surveys:
    post:
      operationId: createSurvey
      tags: [Surveys]
      summary: Create survey
      description: Create a new survey.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyRequest"
      responses:
        "201":
          description: Survey created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listSurveys
      tags: [Surveys]
      summary: List surveys
      description: List surveys for the current organization with filtering and pagination.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, closed]
        - name: search
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Survey list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
  /api/v1/surveys/{surveyId}:
    get:
      operationId: getSurvey
      tags: [Surveys]
      summary: Get survey
      description: Get a survey by ID.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "404":
          description: Survey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Wrong organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateSurvey
      tags: [Surveys]
      summary: Update survey
      description: Update a survey (partial update).
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSurveyRequest"
      responses:
        "200":
          description: Survey updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
    delete:
      operationId: deleteSurvey
      tags: [Surveys]
      summary: Delete survey
      description: Soft-delete a survey.
      parameters:
        $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/surveys/{surveyId}/publish:
    post:
      operationId: publishSurvey
      tags: [Surveys]
      summary: Publish survey
      description: Publish a survey so it can accept responses.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "400":
          description: Survey has no questions or not in draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/surveys/{surveyId}/close:
    post:
      operationId: closeSurvey
      tags: [Surveys]
      summary: Close survey
      description: Close a published survey to stop accepting responses.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
  /api/v1/surveys/{surveyId}/duplicate:
    post:
      operationId: duplicateSurvey
      tags: [Surveys]
      summary: Duplicate survey
      description: Create a copy of a survey.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "201":
          description: Survey duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
...
  /api/v1/public/surveys/{surveyId}:
    get:
      operationId: getPublicSurvey
      tags: [Surveys]
      summary: Get public survey
      description: Get a published survey for public viewing (no auth required).
      security: []
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Public survey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSurvey"
        "404":
          description: Survey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Survey not published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/templates:
    get:
      operationId: listTemplates
      tags: [Templates]
      summary: List templates
      description: List available templates (system + custom).
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Template"
    post:
      operationId: saveAsTemplate
      tags: [Templates]
      summary: Save as template
      description: Save a survey as a custom template.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTemplateRequest"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
...
  /api/v1/surveys/from-template/{templateId}:
    post:
      operationId: createFromTemplate
      tags: [Templates]
      summary: Create from template
      description: Create a new survey from a template.
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Survey created from template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SurveyId:
      name: surveyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: invalid request body
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    CreateSurveyRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
        ui_mode:
          type: string
          enum: [classic, minimal, cards, conversational]
    UpdateSurveyRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ui_mode:
          type: string
          enum: [classic, minimal, cards, conversational]
        status:
          type: string
        structure:
          $ref: "#/components/schemas/SurveyStructure"
        settings:
          $ref: "#/components/schemas/SurveySettings"
        theme:
          $ref: "#/components/schemas/SurveyTheme"
    SurveyStructure:
      type: object
      description: Survey structure defining questions, blocks, and flow.
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionDef"
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/BlockDef"
        flow:
          type: array
          items:
            $ref: "#/components/schemas/FlowStep"
    QuestionDef:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [text, textarea, rating, single_choice, multiple_choice, dropdown, yes_no, date, time, number, email, phone, file_upload, image_choice, ranking, slider]
        required:
          type: boolean
        options:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              label:
                type: string
              value:
                type: string
        settings:
          type: object
          additionalProperties: true
    BlockDef:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        settings:
          type: object
          additionalProperties: true
    FlowStep:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
    SurveySettings:
      type: object
      description: Survey behavior settings.
      properties:
        is_anonymous:
          type: boolean
        allow_multiple_responses:
          type: boolean
        show_progress_bar:
          type: boolean
        show_question_numbers:
          type: boolean
        shuffle_questions:
          type: boolean
        shuffle_options:
          type: boolean
        require_all_questions:
          type: boolean
        show_welcome_screen:
          type: boolean
        welcome_title:
          type: string
        welcome_message:
          type: string
        show_thank_you_screen:
          type: boolean
        thank_you_title:
          type: string
        thank_you_message:
          type: string
        redirect_url:
          type: string
        deadline:
          type: string
          format: date-time
        max_responses:
          type: integer
        estimated_duration:
          type: integer
        background_image:
          type: string
        logo_image:
          type: string
        favicon:
          type: string
        custom_css:
          type: string
        custom_js:
          type: string
    SurveyTheme:
      type: object
      description: Survey visual theme customization.
      properties:
        colors:
          type: object
          properties:
            primary:
              type: string
            secondary:
              type: string
            accent:
              type: string
            background:
              type: string
            text:
              type: string
            button:
              type: string
        font:
          type: object
          properties:
            family:
              type: string
            size:
              type: string
            heading_weight:
              type: string
        border_radius:
          type: integer
        background_image:
          type: string
        background_color:
          type: string
        logo_image:
          type: string
    Survey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, published, closed]
        ui_mode:
          type: string
        structure:
          $ref: "#/components/schemas/SurveyStructure"
        settings:
          $ref: "#/components/schemas/SurveySettings"
        theme:
          $ref: "#/components/schemas/SurveyTheme"
        response_count:
          type: integer
        view_count:
          type: integer
        published_at:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PublicSurvey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
        ui_mode:
          type: string
        structure:
          $ref: "#/components/schemas/SurveyStructure"
        settings:
          $ref: "#/components/schemas/SurveySettings"
        theme:
          $ref: "#/components/schemas/SurveyTheme"
        share_url:
          type: string
          format: uri
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
        next_cursor:
          type: string
          description: Base64-encoded cursor for the next page. Empty string if no more results.
        total:
          type: integer
...
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        structure:
          $ref: "#/components/schemas/SurveyStructure"
        theme:
          $ref: "#/components/schemas/SurveyTheme"
        cover_image_url:
          type: string
          nullable: true
        is_featured:
          type: boolean
        use_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SaveTemplateRequest:
      type: object
      required: [survey_id, name]
      properties:
        survey_id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
...